Polish CLI feature tip terminal transition
